moved quickview modal content to a new component
